chatapp2/requirements.txt
flask==3.0.3
flask-socketio==5.3.6
mysql-connector-python==8.4.0
werkzeug==3.0.3
python-socketio==5.11.3
python-engineio==4.9.1
eventlet==0.36.1
